backend/requirements.txt
beautifulsoup4
lxml
numpy
hnswlib
sentence-transformers
fastapi
uvicorn
tqdm
requests
python-dotenv
